+++ EngineCore/Include/SimpleEngine/Core/Time/GameTime.h
@@ -15,19 +15,19 @@ class GameTime final : public TimeTick
 
 public:
     /** 현재 시간 배율을 반환합니다. (기본값: 1.0) */
-    [[nodiscard]] float GetTimeScale() const { return time_scale; }
+    [[nodiscard]] double GetTimeScale() const { return time_scale; }
 
     /** 일시정지 상태인지 반환합니다. */
     [[nodiscard]] bool IsPaused() const { return paused; }
 
     /** 시간 배율을 설정합니다. */
-    void SetTimeScale(float scale) { time_scale = scale; }
+    void SetTimeScale(double scale) { time_scale = scale; }
 
     /** 일시정지 상태를 설정합니다. */
     void SetPaused(bool pause) { paused = pause; }
 
 private:
-    float time_scale = 1.0f;
+    double time_scale = 1.0;
     bool paused = false;
 };
 } // namespace se
